+++ src/components/RayChart/index.tsx
@@ -36,17 +36,19 @@ import {
   PictorialBarChart,
 } from 'echarts/charts' // 系列类型(后缀都为 `SeriesOption`)
 import { LabelLayout, UniversalTransition } from 'echarts/features' // 标签自动布局, 全局过渡动画等特性
-import { CanvasRenderer } from 'echarts/renderers' // `echarts` 渲染器
+import {
+  CanvasRenderer,
+  // SVGRenderer,
+} from 'echarts/renderers' // `echarts` 渲染器
 
 import { useSetting } from '@/store'
 import { cloneDeep, throttle } from 'lodash-es'
 import { on, off, completeSize } from '@/utils/element'
 import { call } from '@/utils/vue/index'
 import { setupChartTheme, loadingOptions } from './helper'
-import { LAYOUT_CONTENT_REF } from '@/appConfig/routerConfig'
+import { APP_THEME } from '@/appConfig/designConfig'
 
 import type { PropType } from 'vue'
-import type { EChartsInstance } from '@/types/modules/component'
 import type { AnyFC, MaybeArray } from '@/types/modules/utils'
 import type { DebouncedFunc } from 'lodash-es'
 import type {
@@ -54,9 +56,15 @@ import type {
   AutoResize,
   ChartTheme,
 } from '@/components/RayChart/type'
-import type { UseResizeObserverReturn } from '@vueuse/core'
+import type {
+  UseResizeObserverReturn,
+  MaybeComputedElementRef,
+  MaybeElement,
+} from '@vueuse/core'
+import type { ECharts, EChartsCoreOption } from 'echarts/core'
 
 export type EChartsExtensionInstallRegisters = typeof CanvasRenderer
+export type { RayChartInst } from './type'
 
 const RayChart = defineComponent({
   name: 'RayChart',
@@ -116,21 +124,19 @@ const RayChart = defineComponent({
       type: Object as PropType<echarts.EChartsCoreOption>,
       default: () => ({}),
     },
-    success: {
+    onSuccess: {
       /**
        *
        * 返回 chart 实例
        *
        * 渲染成功回调函数
        *
-       * () => EChartsInstance
+       * () => ECharts
        */
-      type: [Function, Array] as PropType<
-        MaybeArray<(e: EChartsInstance) => void>
-      >,
+      type: [Function, Array] as PropType<MaybeArray<(e: ECharts) => void>>,
       default: null,
     },
-    error: {
+    onError: {
       /**
        *
        * 渲染失败回调函数
@@ -148,13 +154,12 @@ const RayChart = defineComponent({
       /**
        *
        * 是否自动跟随模板主题切换
-       *
        * 如果开启此属性, 则会覆盖 `theme` 属性
        *
-       * 注意: 这个属性重度依赖此模板, 所以默认不开启. 并且动态切换主题有一定的性能问题
+       * 注意: 这个属性重度依赖此模板
        */
       type: Boolean,
-      default: false,
+      default: true,
     },
     use: {
       /**
@@ -180,15 +185,37 @@ const RayChart = defineComponent({
       type: Object as PropType<LoadingOptions>,
       default: () => loadingOptions(),
     },
+    observer: {
+      /**
+       *
+       * 需要被监听尺寸的元素
+       * 需要开启 autoResize 才能生效
+       * 默认以父元素作为监听对象
+       */
+      type: Object as PropType<MaybeComputedElementRef<MaybeElement>>,
+      default: null,
+    },
+    throttleWait: {
+      /** 节流等待时间 */
+      type: Number,
+      default: 500,
+    },
+    animation: {
+      /** 是否强制启用渲染动画 */
+      type: Boolean,
+      default: true,
+    },
   },
   setup(props, { expose }) {
     const settingStore = useSetting()
     const { themeValue } = storeToRefs(settingStore)
     const rayChartRef = ref<HTMLElement>() // `echart` 容器实例
-    const echartInstanceRef = ref<EChartsInstance>() // `echart` 拷贝实例, 解决直接使用响应式实例带来的问题
-    let echartInstance: EChartsInstance // `echart` 实例
-    let resizeThrottle: DebouncedFunc<AnyFC> // resize 防抖方法实例
-    let resizeOvserverReturn: UseResizeObserverReturn | undefined
+    const rayChartWrapperRef = ref<HTMLElement>()
+    const echartInstanceRef = ref<ECharts>() // `echart` 实例
+    let echartInstance: ECharts | null // `echart` 拷贝实例, 解决直接使用响应式实例带来的问题
+    let resizeThrottleReturn: DebouncedFunc<AnyFC> | null // resize 防抖方法实例
+    let resizeOvserverReturn: UseResizeObserverReturn | null
+    const { echartTheme } = APP_THEME
 
     const cssVarsRef = computed(() => {
       const cssVars = {
@@ -198,9 +225,6 @@ const RayChart = defineComponent({
 
       return cssVars
     })
-    const modelLoadingOptions = computed(() =>
-      loadingOptions(props.loadingOptions),
-    )
 
     /**
      *
@@ -237,7 +261,7 @@ const RayChart = defineComponent({
       echarts.use([CanvasRenderer]) // 注册渲染器
 
       try {
-        echarts.use(props.use)
+        echarts.use(props.use?.filter(Boolean))
       } catch (e) {
         console.error(
           'Error: wrong property and method passed in extend attribute',
@@ -253,10 +277,17 @@ const RayChart = defineComponent({
      *
      * 如果有需要特殊全局配置的可以在此继续写...
      */
-    const combineChartOptions = () => {
-      let options = cloneDeep(props.options)
+    const combineChartOptions = (ops: EChartsCoreOption) => {
+      let options = cloneDeep(ops)
 
-      const assign = (opts: object) => Object.assign({}, options, opts)
+      const assign = (opts: object) =>
+        Object.assign(
+          {
+            animation: true,
+          },
+          options,
+          opts,
+        )
 
       if (props.showAria) {
         options = assign({
@@ -277,17 +308,16 @@ const RayChart = defineComponent({
      * 渲染 `echart`
      *
      * 缓存两个实例
-     *
      * 直接使用响应式代理实例会出现诡异的问题, 例如 `legend` 点击时报错
      */
-    const renderChart = (theme: ChartTheme = 'macarons') => {
+    const renderChart = (theme: ChartTheme = echartTheme) => {
       /** 获取 dom 容器 */
       const element = rayChartRef.value as HTMLElement
       /** 获取配置项 */
-      const options = combineChartOptions()
+      const options = combineChartOptions(props.options)
       /** 获取 dom 容器实际宽高 */
       const { height, width } = element.getBoundingClientRect()
-      const { success, error } = props
+      const { onSuccess, onError } = props
 
       try {
         /** 注册主题 */
@@ -305,16 +335,22 @@ const RayChart = defineComponent({
         echartInstanceRef.value = echartInstance
 
         /** 设置 options 配置项 */
-        options && echartInstance.setOption(options)
+        options && echartInstance.setOption({})
+
+        if (props.animation) {
+          setTimeout(() => {
+            options && echartInstance?.setOption(options)
+          })
+        }
 
         /** 渲染成功回调 */
-        if (success) {
-          call(success, echartInstance)
+        if (onSuccess) {
+          call(onSuccess, echartInstance)
         }
       } catch (e) {
         /** 渲染失败回调 */
-        if (error) {
-          call(error)
+        if (onError) {
+          call(onError)
         }
 
         console.error('RayChart render error: ', e)
@@ -329,9 +365,9 @@ const RayChart = defineComponent({
      */
     const renderThemeChart = (bool?: boolean) => {
       if (props.autoChangeTheme) {
-        bool ? renderChart('macarons-dark') : renderChart()
+        bool ? renderChart(`${echartTheme}-dark`) : renderChart()
 
-        return void 0
+        return
       }
 
       if (!props.theme) {
@@ -357,10 +393,51 @@ const RayChart = defineComponent({
       }
     }
 
+    const mount = () => {
+      // 避免重复渲染
+      if (echartInstance?.getDom()) {
+        console.warn(
+          'RayChart mount: There is a chart instance already initialized on the dom. Execution was interrupted',
+        )
+
+        return
+      }
+
+      if (props.autoChangeTheme) {
+        /** 注册 echarts */
+        renderThemeChart(themeValue.value)
+      } else {
+        props.theme ? renderChart(`${echartTheme}-dark`) : renderChart()
+      }
+
+      /** 注册事件 */
+      if (props.autoResize) {
+        resizeThrottleReturn = throttle(resizeChart, props.throttleWait)
+        /** 监听内容区域尺寸变化更新 chart */
+        resizeOvserverReturn = useResizeObserver(
+          props.observer || rayChartWrapperRef,
+          resizeThrottleReturn,
+        )
+
+        on(window, 'resize', resizeThrottleReturn)
+      }
+    }
+
+    const unmount = () => {
+      /** 卸载 echarts */
+      destroyChart()
+      /** 卸载事件柄 */
+      resizeThrottleReturn && off(window, 'resize', resizeThrottleReturn)
+      /** 注销防抖 */
+      resizeThrottleReturn?.cancel()
+      /** 注销 observer 监听 */
+      resizeOvserverReturn?.stop?.()
+    }
+
     /** 监听全局主题变化, 然后重新渲染对应主题 echarts */
     watch(
-      () => [themeValue.value],
-      ([theme]) => {
+      () => themeValue.value,
+      (theme) => {
         /**
          *
          * Q: 为什么需要重新卸载再渲染
@@ -375,19 +452,19 @@ const RayChart = defineComponent({
       },
     )
 
+    /**
+     *
+     * 贴花跟随主题渲染
+     *
+     * 自动跟随模板主题或者指定主题皆可
+     */
     watch(
       () => props.showAria,
       () => {
         destroyChart()
 
-        /**
-         *
-         * 贴花跟随主题渲染
-         *
-         * 自动跟随模板主题或者指定主题皆可
-         */
         if (props.autoChangeTheme || props.theme) {
-          themeValue.value ? renderChart('macarons-dark') : renderChart()
+          themeValue.value ? renderChart(`${echartTheme}-dark`) : renderChart()
         } else {
           renderChart()
         }
@@ -397,26 +474,35 @@ const RayChart = defineComponent({
     /** 显示/隐藏加载动画 */
     watch(
       () => props.loading,
-      (newData) => {
-        newData
-          ? echartInstance?.showLoading(modelLoadingOptions.value)
+      (ndata) => {
+        ndata
+          ? echartInstance?.showLoading(props.loadingOptions)
           : echartInstance?.hideLoading()
       },
     )
 
-    /** 监听 options 变化 */
     if (props.watchOptions) {
+      /** 监听 options 变化 */
       watch(
-        () => props.watchOptions,
-        () => {
+        () => props.options,
+        (noptions) => {
           /** 重新组合 options */
-          const options = combineChartOptions()
+          const options = combineChartOptions(noptions)
 
           /** 如果 options 发生变动更新 echarts */
           echartInstance?.setOption(options)
         },
+        {
+          deep: true,
+        },
       )
     }
+
+    expose({
+      echart: echartInstanceRef,
+      dispose: unmount,
+      render: mount,
+    })
 
     onBeforeMount(async () => {
       /** 注册 echarts 组件与渲染器 */
@@ -425,51 +511,25 @@ const RayChart = defineComponent({
 
     onMounted(() => {
       nextTick(() => {
-        /** 注册 echarts */
-        if (props.autoChangeTheme) {
-          renderThemeChart(themeValue.value)
-        } else {
-          props.theme ? renderChart('macarons-dark') : renderChart()
-        }
-
-        /** 注册事件 */
-        if (props.autoResize) {
-          resizeThrottle = throttle(resizeChart, 500)
-
-          on(window, 'resize', resizeThrottle)
-        }
-
-        /** 监听内容区域尺寸变化更新 chart */
-        resizeOvserverReturn = useResizeObserver(
-          LAYOUT_CONTENT_REF.value as unknown as Ref<HTMLElement>,
-          resizeThrottle,
-        )
+        mount()
       })
     })
 
     onBeforeUnmount(() => {
-      /** 卸载 echarts */
-      destroyChart()
-      /** 卸载事件柄 */
-      off(window, 'resize', resizeThrottle)
-      /** 注销防抖 */
-      resizeThrottle.cancel()
-      resizeOvserverReturn?.stop?.()
-    })
-
-    expose({
-      echart: echartInstanceRef,
+      unmount()
     })
 
     return {
       rayChartRef,
       cssVarsRef,
-      echartInstance: echartInstanceRef,
+      rayChartWrapperRef,
     }
   },
   render() {
     return (
-      <div class="ray-chart" style={[this.cssVarsRef]} ref="rayChartRef"></div>
+      <div class="ray-chart" style={[this.cssVarsRef]} ref="rayChartWrapperRef">
+        <div class="ray-chart__container" ref="rayChartRef"></div>
+      </div>
     )
   },
 })