+++ src/components/RayIframe/index.ts
@@ -1,3 +1,6 @@
 import RayIframe from './src/index'
 
+import type { RayIframeInst } from './src/index'
+
 export default RayIframe
+export type { RayIframeInst }