+++ src/router/routes.ts
@@ -3,8 +3,6 @@ import { getAppRawRoutes } from './routeModules'
 import { ROOT_ROUTE } from '@/appConfig/appConfig'
 import { expandRoutes } from '@/router/helper/expandRoutes'
 
-const { path } = ROOT_ROUTE
-
 export default async () => [
   {
     path: '/',
@@ -14,7 +12,7 @@ export default async () => [
   {
     path: '/',
     name: 'layout',
-    redirect: path,
+    redirect: ROOT_ROUTE.path,
     component: Layout,
     children: expandRoutes(getAppRawRoutes()),
   },