+++ src/types/modules/component.ts
@@ -1,9 +1,6 @@
-import type { ECharts } from 'echarts/core'
 import type { MenuOption, MenuDividerOption, MenuGroupOption } from 'naive-ui'
 
 export type ComponentSize = 'small' | 'medium' | 'large'
-
-export type EChartsInstance = ECharts
 
 export type Placement = 'top' | 'right' | 'bottom' | 'left'